+++ src/js/loaders/MainLoader.js
@@ -3,7 +3,6 @@ import "./FeatureChecker"
 import "./Settings"
 import "./Navigation"
 import "./Listeners"
-import "./DBControllers"
 import Report from "../main/report"
 
 import LoadState from "../services/LoadState"