+++ src/js/ui/DOM/Library/settings/SettingsSectionElement.js
@@ -1,14 +1,8 @@
 import DOM from "../../Classes/dom"
 import Title from "../object/title"
-import FieldsContainer from "../../../../tools/validation/fieldsContainer"
-import FieldChecker from "../../../../tools/validation/fieldChecker"
 
 export default class SettingsSectionElement {
     constructor(options, object) {
-        new FieldsContainer([[], {
-            name: new FieldChecker({ type: "string" }),
-        }]).set(options)
-
         return new DOM({
             new: "div",
             class: ["settings-section-chunk"],