+++ src/js/ui/DOM/Library/style/padding.js
@@ -1,12 +1,13 @@
 import DOM from "../../Classes/dom"
 
 export default class Padding {
-    constructor(content, padding) {
+    constructor(content, padding, style = {}) {
         padding = padding.toString()
         padding = (padding.match(/\d$/) ? `${padding}px` : padding)
         return new DOM({
             new: "div",
             style: {
+                ...style,
                 padding,
             },
             content,